test: seperate length tests
